+++ backend/canisters/user_index/api/src/main.rs
@@ -68,7 +68,6 @@ fn main() {
     generate_ts_method!(user_index, set_user_upgrade_concurrency);
     generate_ts_method!(user_index, set_moderation_flags);
     generate_ts_method!(user_index, set_username);
-    generate_ts_method!(user_index, submit_proof_of_unique_personhood);
     generate_ts_method!(user_index, suspend_user);
     generate_ts_method!(user_index, unsuspend_user);
     generate_ts_method!(user_index, update_bot);