+++ backend/canisters/user_index/impl/src/jobs/mod.rs
@@ -1,6 +1,7 @@
 use crate::RuntimeState;
 pub mod make_pending_payments;
 pub mod remove_from_online_users_canister;
+pub mod remove_lapsed_unique_person_proofs;
 pub mod reset_leaderboard;
 pub mod submit_message_to_modclub;
 pub mod sync_events_to_local_user_index_canisters;
@@ -10,6 +11,7 @@ pub mod upgrade_canisters;
 pub(crate) fn start(state: &RuntimeState) {
     make_pending_payments::start_job_if_required(state);
     remove_from_online_users_canister::start_job_if_required(state);
+    remove_lapsed_unique_person_proofs::start_job_if_required(state);
     submit_message_to_modclub::start_job_if_required(state);
     sync_events_to_local_user_index_canisters::start_job_if_required(state);
     sync_users_to_identity_canister::start_job_if_required(state);