+++ backend/canisters/user_index/impl/src/lifecycle/inspect_message.rs
@@ -22,7 +22,6 @@ fn accept_if_valid(state: &RuntimeState) {
         | "set_hide_online_status"
         | "set_moderation_flags"
         | "set_username"
-        | "submit_proof_of_unique_personhood"
         | "update_bot"
         | "update_diamond_membership_subscription" => state.is_caller_openchat_user(),
         "suspend_user" | "unsuspend_user" => state.is_caller_platform_moderator(),