+++ backend/canisters/user_index/impl/src/model/user_map.rs
@@ -796,6 +796,16 @@ impl UserMap {
         }
     }
 
+    pub fn remove_proof_of_unique_personhood(&mut self, user_id: UserId, now: TimestampMillis) -> bool {
+        if let Some(user) = self.users.get_mut(&user_id) {
+            if user.unique_person_proof.take().is_some() {
+                user.date_updated = now;
+                return true;
+            }
+        }
+        false
+    }
+
     pub fn unique_person_proofs_submitted(&self) -> u32 {
         self.unique_person_proofs_submitted
     }